+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-02 17:15:56
-// Commit: docs(core): update readme and API contracts for core
+// Updated at: 2026-07-03 11:05:08
+// Commit: chore(core): update dependencies and cleanup unused imports in core
 
 class coreComponentState {
-  final String commitRef = '2480';
-  final String lastMessage = 'docs(core): update readme and API contracts for core';
-  final DateTime timestamp = DateTime.parse('2026-07-02T17:15:56+05:00');
+  final String commitRef = '2512';
+  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
+  final DateTime timestamp = DateTime.parse('2026-07-03T11:05:08+05:00');
 }